fix no-arg command
